+++ src/main/java/kr/codesqaud/cafe/controller/dto/user/JoinDTO.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.controller.dto;
+package kr.codesqaud.cafe.controller.dto.user;
 
 import kr.codesqaud.cafe.domain.User;
